requirements.txt
numpy
scikit-learn
umap-learn
matplotlib
seaborn
pandas
plotly
